% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ab.fam.set.combns.by.pool.R
\docType{data}
\name{ab.fam.set.combns.by.pool}
\alias{ab.fam.set.combns.by.pool}
\title{Family set combination by pool data - 'pooling for individual parentage assignment' example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
ab.fam.set.combns.by.pool
}
\description{
Identifies family set combinations with pooled samples - simulated data (refer to Hamilton et al. in prep for details)
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
