% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ab.fam.set.combns.R
\docType{data}
\name{ab.fam.set.combns}
\alias{ab.fam.set.combns}
\title{Family set combination data - 'pooling for individual parentage assignment' example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
ab.fam.set.combns
}
\description{
Identifies family set combinations, family sets within combinations and families with family sets - simulated data (refer to Hamilton et al. in prep for details)
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
