% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ab.snp.dat.indiv.R
\docType{data}
\name{ab.snp.dat.indiv}
\alias{ab.snp.dat.indiv}
\title{SNP data for individuals - 'pooling for individual parentage assignment' example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
ab.snp.dat.indiv
}
\description{
Individual SNP intensity and genotype call data - simulated data (refer to Hamilton et al. in prep for details)
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
