% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ham.fam.set.combns.by.pool.R
\docType{data}
\name{Ham.fam.set.combns.by.pool}
\alias{Ham.fam.set.combns.by.pool}
\title{Family set combination by pool data for small worked example in Hamilton et al. in prep}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
Ham.fam.set.combns.by.pool
}
\description{
Identifies family set combinations with pooled samples
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
