% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ham.fam.set.combns.R
\docType{data}
\name{Ham.fam.set.combns}
\alias{Ham.fam.set.combns}
\title{Family set combination data for small worked example in Hamilton et al. in prep}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
Ham.fam.set.combns
}
\description{
Identifies family set combinations, family sets within combinations and families with family sets
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
