% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ham.fams.R
\docType{data}
\name{Ham.fams}
\alias{Ham.fams}
\title{Family data for small worked example in Hamilton et al. in prep}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
Ham.fams
}
\description{
Identifies families and their sires and dams
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
