% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ham.snp.dat.indiv.R
\docType{data}
\name{Ham.snp.dat.indiv}
\alias{Ham.snp.dat.indiv}
\title{Individual data for small worked example in Hamilton et al. in prep}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
Ham.snp.dat.indiv
}
\description{
Individual SNP intensity and genotype call data
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
