% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ham.snp.dat.pools.R
\docType{data}
\name{Ham.snp.dat.pools}
\alias{Ham.snp.dat.pools}
\title{Data for DNA pools from small worked example in Hamilton et al. in prep}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
Ham.snp.dat.pools
}
\description{
SNP intensity and genotype call data for DNA pools
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
