% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shrimp.fams.R
\docType{data}
\name{shrimp.fams}
\alias{shrimp.fams}
\title{Family data - 'pooling by phenotype' example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
shrimp.fams
}
\description{
Identifies families and their sires and dams - simulated data (refer to Hamilton et al. in prep for details)
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
