% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrimp.map.R
\docType{data}
\name{shrimp.map}
\alias{shrimp.map}
\title{Map data - 'pooling by phenotype' simulation example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
shrimp.fams
}
\description{
Genetic map file for each SNP - used for simulations presented in Hamilton et al. in prep
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
