% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrimp.true.snp.param.indiv.R
\docType{data}
\name{shrimp.true.snp.param.indiv}
\alias{shrimp.true.snp.param.indiv}
\title{True SNP parameter data - 'pooling by phenotype' simulation example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
shrimp.fams
}
\description{
Assumed 'true' values for SNP parameters - used for simulations presented in Hamilton et al. in prep
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
