% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrimp.sim.fam.sets.R
\docType{data}
\name{sim.fam.sets}
\alias{sim.fam.sets}
\title{Assignment of families to family sets - 'pooling by phenotype' simulation example}
\format{
Data frame
}
\source{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\usage{
shrimp.fams
}
\description{
Assigns families to family sets - used for simulations presented in Hamilton et al. in prep
}
\references{
Hamilton et al. (in prep) Methods of parentage assignment for pooled samples using low-density SNP data
}
\author{
Matthew Hamilton <matthew.hamilton@csiro.au>
}
\keyword{datasets}
