% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.param.indiv.fun.R
\name{snp.param.indiv.fun}
\alias{snp.param.indiv.fun}
\title{snp.param.indiv.fun}
\usage{
snp.param.indiv.fun(
  snp.dat.indiv,
  min.count = 2,
  min.intensity = 0,
  gen.plots = FALSE
)
}
\arguments{
\item{snp.dat.indiv}{is a data frame with the following headings (class in parentheses):
\itemize{
 \item{'SAMPLE_ID' is the sample identifier.  Samples must be from diploid individuals (i.e. not pools) (integer).}
 \item{'SNP_ID' is the SNP identifier (character).} 
 \item{'INTENSITY_A' is the area/intensity for allele A (numeric).}
 \item{'INTENSITY_B' is the area/intensity for allele B (numeric).} 
 \item{'A_ALLELE' is the base represented by allele A (i.e. 'A', 'C', 'G' or 'T') (character).}
 \item{'B_ALLELE' is the base represented by allele B (i.e. 'A', 'C', 'G' or 'T') (character).}
 \item{'GENOTYPE' is the SNP genotype call (e.g. 'AT', 'TT').  NA if missing (character).}
}}

\item{min.count}{is an integer (default = 2) representing the minimum number of each genotype (i.e. 'AA', 'AB', 'BB') for each SNP
required for the computation of SNP parameters.}

\item{min.intensity}{is a numeric variable (default = 0).  If the square root of the sum of INTENSITY_A squared and 
INTENSITY_B squared is less than min.intensity then this record is excluded in the computation of SNP parameters.
That is, observations that fall into an arc with a radius equal to min.intensity in the lower left of
signal intensity scatter plots are excluded.}

\item{gen.plots}{is a logical variable.  If TRUE then signal intensity scatter plots and 
signal intensity against allelic proportion scatter plots for each SNP are generated and saved to disc}
}
\value{
1. A data frame containing estimates of SNP sepecific parameters (refer to Henshall et al. 2014): 
\itemize{
 \item{'SNP_ID' is the SNP identifier.} 
 \item{'N_AA' is the count of homozygous A (AA) genotypes.}
 \item{'MEAN_P_AA' is the mean of allelic proportion for homozygous A genotypes.}
 \item{'SD_P_AA' is the standard deviation of allelic proportion for homozygous A genotypes.} 
 \item{'N_AB' is the count of heterozygous (AB) genotypes.}
 \item{'MEAN_P_AB' is the mean of allelic proportion for heterozygous (AB) genotypes.}
 \item{'SD_P_AB' is the standard deviation of allelic proportion for heterozygous (AB) genotypes.}  
 \item{'N_BB' is the count of homozygous B (BB) genotypes.}
 \item{'MEAN_P_BB' is the mean of allelic proportion for homozygous B genotypes.}
 \item{'SD_P_BB' is the standard deviation of allelic proportion for homozygous B genotypes.} 
 \item{'WELCH_A' is the welsh statistic for the interval between MEAN_P_AA and MEAN_P_AB.}
 \item{'WELCH_B' is the welsh statistic for the interval between MEAN_P_AB and MEAN_P_BB.}
 \item{'A_ALLELE_FREQ' is the A allele frequency computed from genotype counts.}
 \item{'B_ALLELE_FREQ' is the B allele frequency computed from genotype counts.}
 \item{'A_ALLELE' is the base represented by allele A (i.e. 'A', 'C', 'G' or 'T').}
 \item{'B_ALLELE' is the base represented by allele B (i.e. 'A', 'C', 'G' or 'T').}
}

2. '[SNP_ID].intensity.png' (if gen.plots = TRUE): signal intensity scatter plots saved in a sub-directory named 'Results/SNP_intensity_plots' 
in the working directory.

3. '[SNP_ID].prop.png' (if gen.plots = TRUE): signal intensity against allelic proportion scatter plots saved in a sub-directory named directory named
'Results/SNP_allelic_prop_plots' in the working directory. See Figure 1 of Henshall et al. 2014
}
\description{
This function generates SNP parameters (specifically allelic proportion means, standard deviations and welch statistics) 
as described in the 'Estimation of SNP specific parameters' section of Henshall et al. (2014).  
If specified, it also generates signal intensity scatter plots and signal intensity against allelic proportion scatter plots 
(e.g. Figure 1 of Henshall et al. 2014) for each SNP.
}
\examples{
#Retrieve data for small worked example from Hamilton 2020
data(Ham.snp.dat.indiv)

#Compute SNP parameters
snp.param.indiv.fun(Ham.snp.dat.indiv)
}
\references{
Henshall JM, Dierens, L Sellars MJ (2014) Quantitative analysis of low-density SNP data for parentage assignment and estimation of family contributions to pooled samples. Genetics Selection Evolution 46, 51. https://doi 10.1186/s12711-014-0051-y

Hamilton MG (2020) Maximum likelihood parentage assignment using quantitative genotypes
}
