% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.param.pools.fun.R
\name{snp.param.pools.fun}
\alias{snp.param.pools.fun}
\title{snp.param.pools.fun}
\usage{
snp.param.pools.fun(snp.param.indiv, n.in.pools)
}
\arguments{
\item{snp.param.indiv}{is the output of snp.param.indiv.fun.  That is, it is a data frame with the following headings (class in parentheses):
\itemize{
 \item{'SNP_ID' is the SNP identifier (character).} 
 \item{'N_AA' is the count of homozygous A (AA) genotypes (integer).}
 \item{'MEAN_P_AA' is the mean of allelic proportion for homozygous A genotypes (numeric).}
 \item{'SD_P_AA' is the standard deviation of allelic proportion for homozygous A genotypes (numeric).} 
 \item{'N_AB' is the count of heterozygous (AB) genotypes (integer).}
 \item{'MEAN_P_AB' is the mean of allelic proportion for heterozygous (AB) genotypes (numeric).}
 \item{'SD_P_AB' is the standard deviation of allelic proportion for heterozygous (AB) genotypes (numeric).}  
 \item{'N_BB' is the count of homozygous B (BB) genotypes (integer).}
 \item{'MEAN_P_BB' is the mean of allelic proportion for homozygous B genotypes (numeric).}
 \item{'SD_P_BB' is the standard deviation of allelic proportion for homozygous B genotypes (numeric).} 
 \item{'WELCH_A' is the welsh statistic for the interval between MEAN_P_AA and MEAN_P_AB (numeric).}
 \item{'WELCH_B' is the welsh statistic for the interval between MEAN_P_AB and MEAN_P_BB (numeric).}
 \item{'A_ALLELE_FREQ' is the A allele frequency computed from genotype counts (numeric).}
 \item{'B_ALLELE_FREQ' is the B allele frequency computed from genotype counts (numeric).}
 \item{'A_ALLELE' is the base represented by allele A (i.e. 'A', 'C', 'G' or 'T') (character).}
 \item{'B_ALLELE' is the base represented by allele B (i.e. 'A', 'C', 'G' or 'T') (character).}
}}

\item{n.in.pools}{is an integer representing the number of individuals in pools.}
}
\value{
1. A data frame containing estimates of SNP sepecific parameters for pooled genotypes (refer to Hamilton 2020).  
The number of columns in the data frame is dependent on the number of possible pooled genotypes which is determined from n.in.pools. 
The following is an example of output where n.in.pools = 2. 
\itemize{
 \item{'SNP_ID' is the SNP identifier.} 
 \item{'MEAN_P_AAAA' is the mean of allelic proportion for homozygous A genotypes.}
 \item{'SD_P_AAAA' is the standard deviation of allelic proportion for homozygous A genotypes.} 
 \item{'MEAN_P_AAAB' is the mean of allelic proportion for unordered AAAB genotypes.}
 \item{'SD_P_AAAB' is the standard deviation of allelic proportion for unordered AAAB genotypes.} 
 \item{'MEAN_P_AABB' is the mean of allelic proportion for unordered AABB genotypes.}
 \item{'SD_P_AABB' is the standard deviation of allelic proportion for unordered AABB genotypes.}  
 \item{'MEAN_P_ABBB' is the mean of allelic proportion for unordered ABBB genotypes.}
 \item{'SD_P_ABBB' is the standard deviation of allelic proportion for unordered ABBB genotypes.}
 \item{'MEAN_P_BBBB' is the mean of allelic proportion for homozygous B genotypes.}
 \item{'SD_P_BBBB' is the standard deviation of allelic proportion for homozygous B genotypes.}   
 \item{'A_ALLELE' is the base represented by allele A (i.e. 'A', 'C', 'G' or 'T').}
 \item{'B_ALLELE' is the base represented by allele B (i.e. 'A', 'C', 'G' or 'T').}
}
}
\description{
This function generates SNP parameters (specifically allelic proportion means and standard deviations) 
for pooled genotypes as described in Hamilton et al. (in prep).
}
\examples{
#Retrieve data for small worked example from Hamilton 2020
data(Ham.snp.dat.indiv)

#Compute SNP parameters
Ham.snp.param.indiv <- snp.param.indiv.fun(Ham.snp.dat.indiv)
snp.param.pools.fun(Ham.snp.param.indiv, n.in.pools = 2)
}
\references{
Hamilton MG (2020) Maximum likelihood parentage assignment using quantitative genotypes
}
